refactor hideOrShow fucntion to be a module, import hideOrShow module as showClickedSection
